Added possibility to hide Notifications.
