+++ app/src/main/java/de/mytfg/app/android/gcm/GcmManager.java
@@ -1,9 +1,21 @@
 package de.mytfg.app.android.gcm;
 
+import android.app.NotificationManager;
+import android.app.PendingIntent;
+import android.content.Context;
+import android.content.Intent;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.media.RingtoneManager;
+import android.net.Uri;
 import android.os.Bundle;
+import android.support.v4.app.NotificationCompat;
+import android.util.Log;
 
 import java.util.Hashtable;
 
+import de.mytfg.app.android.R;
+import de.mytfg.app.android.slidemenu.MainActivity;
 import de.mytfg.app.android.utils.BiMap;
 
 /**
@@ -17,6 +29,7 @@ public class GcmManager {
     public GcmManager () {
         notificationIds = new BiMap<>();
         clickCallbacks = new Hashtable<>();
+        receiveCallbacks = new Hashtable<>();
     }
 
     public void setClickCallback(String type, GcmCallback callback) {
@@ -27,7 +40,18 @@ public class GcmManager {
         receiveCallbacks.put(type, callback);
     }
 
-    public int notify(String from, Bundle data) {
+    public void hide(GcmNotification notification) {
+        Integer id = notificationIds.getKey(notification);
+
+        Log.d("GCM", "I should hide");
+        if (id != null) {
+            Log.d("GCM", "ID: " + id);
+            NotificationManager notificationManager = (NotificationManager) MainActivity.context.getSystemService(Context.NOTIFICATION_SERVICE);
+            notificationManager.cancel(id);
+        }
+    }
+
+    public void notify(String from, Bundle data) {
         GcmNotification notification = new GcmNotification(data);
 
         int id;
@@ -40,13 +64,13 @@ public class GcmManager {
             notificationIds.add(id, notification);
         }
 
+        this.sendNotification(data.getString("message"), data.getString("title"), id);
+
         // Call on receive callback if existing
         GcmCallback toCall = receiveCallbacks.get(notification.getType());
         if (toCall != null) {
             toCall.callback(notification);
         }
-
-        return id;
     }
 
     public void clicked(int id) {
@@ -60,6 +84,33 @@ public class GcmManager {
         }
     }
 
+    private void sendNotification(String message, String title, int id) {
 
+        Intent intent = new Intent(MainActivity.context, MainActivity.class);
+        intent.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
+        intent.putExtra("notificationId", id);
+        PendingIntent pendingIntent = PendingIntent.getActivity(MainActivity.context, 0 /* Request code */, intent,
+                PendingIntent.FLAG_UPDATE_CURRENT);
+
+        Uri defaultSoundUri= RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+
+        Bitmap bm = BitmapFactory.decodeResource(MainActivity.context.getResources(), R.mipmap.ic_launcher);
+
+        NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(MainActivity.context)
+                .setLargeIcon(bm)
+                .setSmallIcon(R.mipmap.ic_launcher)
+                .setContentTitle(title)
+                .setContentText(message)
+                .setAutoCancel(true)
+                .setStyle(new NotificationCompat.BigTextStyle()
+                        .bigText(message))
+                .setSound(defaultSoundUri)
+                .setContentIntent(pendingIntent);
+
+        NotificationManager notificationManager =
+                (NotificationManager) MainActivity.context.getSystemService(Context.NOTIFICATION_SERVICE);
+
+        notificationManager.notify(id, notificationBuilder.build());
+    }
 
 }
